fix: rename to plot_tsne_trajectory
